requirements.txt
absl-py==0.5.0
adal==1.2.0
alabaster==0.7.11
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.16
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.37
azure-eventgrid==1.2.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.0.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.3.1
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.3.0
azure-mgmt-containerregistry==2.4.0
azure-mgmt-containerservice==4.2.2
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.2.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.2.1
azure-mgmt-notificationhubs==2.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.5.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.0.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.0.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.3.1
azure-storage-common==1.3.0
azure-storage-file==1.3.1
azure-storage-nspkg==3.0.0
azure-storage-queue==1.3.0
Babel==2.6.0
backcall==0.1.0
bleach==3.0.0
boto==2.49.0
boto3==1.9.22
botocore==1.12.22
bz2file==0.98
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
colorama==0.3.9
cryptography==2.3.1
cycler==0.10.0
cymem==1.31.2
Cython==0.28.5
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
ffmpeg==1.4
ffprobe==0.5
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-SocketIO==3.0.1
Flask-WTF==0.14.2
funcy==1.11
future==0.16.0
gast==0.2.0
gensim==3.6.0
grpcio==1.15.0
gunicorn==19.9.0
h5py==2.7.0
html5lib==1.0.1
idna==2.7
imagesize==1.1.0
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.11
jsonify==0.5
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.0.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==15.1.0
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.4.1
msrest==0.6.1
msrestazure==0.5.1
murmurhash==0.28.0
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.7.0
numexpr==2.6.8
numpy==1.15.2
numpydoc==0.8.0
oauthlib==2.1.0
packaging==18.0
pandas==0.23.4
pandas-datareader==0.7.0
pandasql==0.7.3
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
plac==0.9.6
pluggy==0.7.1
preshed==1.0.1
prometheus-client==0.4.0
prompt-toolkit==2.0.5
protobuf==3.6.1
psutil==5.4.7
py==1.6.0
pycodestyle==2.4.0
pycparser==2.19
pydub==0.23.0
pyflakes==2.0.0
Pygments==2.2.0
PyJWT==1.6.4
pyLDAvis==2.1.2
pylint==2.1.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.2
PyQt5==5.9.2
pyreadline==2.1
PySocks==1.6.8
pytest==3.8.2
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.1
qtconsole==4.4.1
QtPy==1.5.1
regex==2017.4.5
requests==2.19.1
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.0.0
rope==0.11.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
smart-open==1.7.1
snowballstemmer==1.2.1
spacy==2.0.12
SpeechRecognition==3.8.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.12
statsmodels==0.9.0
tables==3.4.4
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.3
thinc==6.10.3
toolz==0.9.0
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
wordcloud==1.5.0
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
